requirements.txt
selenium==4.11.2
pytest==8.3.4
assertpy==1.1
requests==2.32.3
allure-pytest==2.13.5
webdriver-manager==4.0.2